Fix sensor test file name.
